# cmd/api/docs/swagger.yaml
definitions:
...
  models.AdminLogin:
    properties:
      email:
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.AdminSignUp:
    properties:
      confirmpassword:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - confirmpassword
    - email
    - name
    - password
    type: object
  models.CategoryUpdate:
    properties:
      category:
        type: string
    required:
    - category
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/adminlogin:
    post:
      consumes:
      - application/json
      description: Login handler for admin
      parameters:
      - description: Admin login details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin Login
      tags:
      - Admin Authentication
  /admin/categories:
    get:
      consumes:
      - application/json
      description: Display Category details on the admin side
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Category Details to admin side
      tags:
      - Admin Category Management
  /admin/categories/add_category:
    post:
      consumes:
      - application/json
      description: Add a new Category So that movie of that category can be added
      parameters:
      - description: Update Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.CategoryUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add a new Categorys ( Category )
      tags:
      - Admin Category Management
  /admin/categories/delete_category/{id}:
    post:
      consumes:
      - application/json
      description: Delete Category for existing films and delete the films along with
        it
      parameters:
      - description: category-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Category ( Category )
      tags:
      - Admin Category Management
  /admin/createadmin:
    post:
      consumes:
      - application/json
      description: Signup handler for admin
      parameters:
      - description: Admin login details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Admin Signup
      tags:
      - Admin Authentication
  /admin/dashboard:
    get:
      consumes:
      - application/json
      description: Get Amin Home Page with Complete Details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Admin Dashboard
      tags:
      - Admin Dash Board
...
  /admin/sales-report/{period}:
    get:
      consumes:
      - application/json
      description: Get Filtered sales report by week, month and year
      parameters:
      - description: sales report
        in: path
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Filtered Sales Report
      tags:
      - Admin Dash Board
  /admin/users/{page}:
    get:
      consumes:
      - application/json
      description: Retrieve users with pagination to admin side
      parameters:
      - description: Page number
        in: path
        name: page
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get Users Details To Admin
      tags:
      - Admin User Management
  /admin/users/block-users/{id}:
    post:
      consumes:
      - application/json
      description: Block an existing user using user id
      parameters:
      - description: user-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Block  user
      tags:
      - Admin User Management
  /admin/users/unblock-users/{id}:
    post:
      consumes:
      - application/json
      description: Unblock an already blocked user using user id
      parameters:
      - description: user-id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Unblock  User
      tags:
      - Admin User Management
...
swagger: "2.0"